requirements.txt
minecraft_launcher_lib
PyYAML
rich
simple-term-menu
Faker
lxml
